Separate public API from implementation
Major reorganization that establishes proper module boundaries and
improves compilation times by separating public interface from internal
implementation details.

* include/blot/blot.hpp:
  (annotation_options, annotation_result, annotate): Public interface only
  (detail namespace, internal functions): Moved to blot.cpp
* src/libblot/blot.cpp: New file.
* src/libblot/linespan.hpp: Moved from include/blot/
* src/libblot/logger.hpp: Moved from include/blot/
* src/blot/options.hpp (file_options): Moved here.
* src/blot/main.cpp: Updated include paths and formatting
* src/libblot/{assembly,ccj}.cpp: Updated include paths for moved headers
* test/test_blot.cpp: Maintains compatibility with public API
