+++ src/libblot/linespan.hpp
