+++ src/libblot/logger.hpp
